Fix bug where wrong select option would get selected.
This fixes the bug with select elements reported in #107.

The main problem was looping through children of toEl when determining
the correct selectedIndex, since nodes from toEl can get moved to fromEl
while morphing in some cases, so by the time the select handler is
invoked, the toEl may already be missing some of its original children.

On the other hand, at the time when the select handler is invoked,
fromEl's children are already fully patched and in place.