Fix orphaned nodes when matched keyed node moved to new parent
This was causing nodes to not be added back to the from tree when:

1) a non-keyed root parent was removed and re-added
(such as when a sibling above caused a shift in children)
2) when a keyed child of the recreated parent returned
false from onBeforeElUpdated

In such cases, we would not add the children of the skipped update
back to the tree. Additionally, we would invoke both the updated
callbacks and the added callbacks, for the same nodes. This fixes
the scenario by propagating the movedToNewParent state, and when we
find we are skipping an update, but we moved to a new parent, we mark
the to tree equal to the from tree, then continue the traversal,
to ensure we added proper nodes back to the tree.