Perform diff when `toNode` is a DocumentFragment
Presently a diff is not performed when `toNode` is a DocumentFragment — morphdom completely replaces the old element with the new element, even when the element/children have not changed. This behavior can cause unexpected results, such as event listeners disappearing even though the markup for the event trigger hasn't changed.

The fix works by placing the DocumentFragment's contents inside a div, and then using the div's `innerHTML` as the string against which the diff should be applied.