fix keyed lookup when using a cloned node in onBeforeElUpdated
When returning a cloned element in onBeforeElUpdated, the keyed nodes
would still point to the original tree. If morphdom then tries to remove
an element it actually removed it from the DOM instead of removing it
from the cloned node. This is only ever a problem if a cloned node is
returned in onBeforeElUpdated, which is something that Phoenix LiveView
does sometimes.